common: Add support to multiple logger IDs on beerocks_logging
Since the ap_manager thread and the monitor thread will be on the same process, we will
need separate log files for each one.
Currently the beerocks_logging support single logger ID which is "default".
To support custom logger ID add an additional logger_id argument to beerocks_logging
constructor and add logger_id member to the logging class that will be used to configure
the logger instead of "default" logger ID.

On the way, make the module_name on the constructor to be mandatory instead of optional, since every process already passes this argument.
Also, change the logging class string argument to be a const reference.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>