common: Add custom logger ID helper file
When using LOG() macro it is always print to default logger ID which is "default" string.
When we want to have to log files on a single process we have to create a different
logger ID for the non-default logger.

We can't use LOG() macro if we want to print to non-default logger ID.
To avoid replacing all LOG() calls manually on some file, add a helper file that allows with
a single #define and #include at the beginning of a file, do the replacing automatically.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>