common: Add an option to set thread name on beerocks logging
When using Syslog, easylogging does not write the prints to log files, it sends it to Syslog.
On Syslog, it is possible to distinguish prints source by process name.
So in our case, it will be  "beerocks_fronthaul_wlanX", and we need a way to separate between
ap_manager and monitor threads.
To do it, add an option to set the thread name, on logging class, and set it from ap_manager
and monitor thread.
After the change the log will look like this on Syslog:
beerocks_fronthaul_wlan0: <ap_manager> ap_manager_thread.cpp[205] --> waiting to attach to wlan0
beerocks_fronthaul_wlan0: <monitor> monitor_thread.cpp[205] --> waiting to attach to wlan0

The thread name between <> brackets will be used to split the messages from the main log file
to the right prplMesh log file.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>