agent: slave: channel selection: bail out if invalid channel is given.
In certification test 4.5.2, one of the tests is to trigger the
controller to send a channel selection request which contains invalid
channels. Concretely, one of the values is: {0x73 0x03 {0x95 0x99 0x9D}
0x00} i.e. operating class 115, channels 149, 153, 157, preference value
0. Since channels 149, 153, 157 are in operating class 124, not 115,
this is invalid. The test requires an error
(DECLINE_VIOLATES_MOST_RECENTLY_REPORTED_PREFERENCES) in this case.

Refactor the code to report this error. Check the channel against the
operating class's channel list in
channel_selection_get_channel_preference(). Rewrite the
handle_channel_selection_request() function to first analyze the data
and determine the response code and channel to switch to, then send the
response. While we're at it, log a message right away then the channel
has been chosen.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>