test_flows: channel_selection: add test of invalid channel
In certification test 4.5.2, one of the tests is to trigger the
controller to send a channel selection request which contains invalid
channels. Concretely, one of the values is: {0x73 0x03 {0x95 0x99 0x9D}
0x00} i.e. operating class 115, channels 149, 153, 157, preference value
0. Since channels 149, 153, 157 are in operating class 124, not 115,
this is invalid. The test requires an error
(DECLINE_VIOLATES_MOST_RECENTLY_REPORTED_PREFERENCES) in this case.

Add a test to test_flows to validate this case. A very much simplified
request is sent with just a single operating class 82 (which contains
only channel 14) and specifying channel 1. Check the response code. Like
for all channel selection tests, we need to ignore the double response.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>