agent: backhaul_manager: forward AP Metrics Response if mid==0
AP Metrics Response message is created in monitor thread when channel
utilization crosses threshold and then sent to slave thread, which in
turn sends it to backhaul manager thread. The message id of such message
is always 0.

Since this message is complete, backhaul manager must forward it "as
is" to the controller.

On the other hand, if AP Metrics Response message contains a message id
other than 0, then it corresponds to a previously received and forwarded
AP Metrics Query.
In that case, backhaul manager must wait until responses from all slaves
have been received before building and sending an AP Metrics Response to
the controller.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>